shm_free implemented

+++ gbtcp/shm.h
@@ -10,9 +10,9 @@ void shm_deinit();
 void shm_detach();
 
 int shm_malloc(void **, size_t);
-int shm_realloc(void **, int);
+int shm_realloc(void **, size_t);
 void shm_free(void *);
-int shm_alloc_pages(void **, int, int);
-void shm_free_pages(void *, int);
+int shm_alloc_pages(void **, size_t, size_t);
+void shm_free_pages(void *, size_t);
 
 #endif // GBTCP_SHM_H
